+++ src/pwshBedrock/Public/Reset-ModelTally.ps1
@@ -87,6 +87,7 @@ function Reset-ModelTally {
             'mistral.mistral-small-2402-v1:0',
             'mistral.mistral-large-2402-v1:0',
             'mistral.mistral-large-2407-v1:0',
+            'mistral.pixtral-large-2502-v1:0',
             'mistral.mixtral-8x7b-instruct-v0:1',
             'stability.stable-diffusion-xl-v1',
             'stability.stable-image-ultra-v1:0',